feat: update routing for login pages
