+++ layouts/default.vue
@@ -20,6 +20,7 @@ import TopBar from "~/components/app/top-bar.vue";
 
 export default defineComponent({
   components: {AdminSidebar, TopBar},
+  middleware: ['isAuthenticated'],
   setup() {
     const store = useStore<RootStoreState>();
