+++ middleware/isUnauthenticated.ts
@@ -2,6 +2,6 @@ import {defineNuxtMiddleware} from '@nuxtjs/composition-api';
 
 export default defineNuxtMiddleware(({store, redirect}) => {
   if (store.getters['employee/isEmployeeLoggedIn']) {
-    return redirect('/records');
+    return redirect('/');
   }
 });
