+++ pages/timesheets/_employee_id/_start_timestamp.vue
@@ -1,4 +1,4 @@
 <script>
-import RecordsPage from "~/pages/records/index.vue";
+import RecordsPage from "~/pages/index.vue";
 export default RecordsPage;
 </script>
